+++ db/db_test_util.cc
@@ -197,7 +197,7 @@ bool DBTestBase::ChangeOptions(int skip_mask) {
 
 #if 1  // scaf new version
 // Switch between different option styles.  Return false if no more to try
-bool DBTestBase::CycleThroughOptions(std::vector<int>& optioncycle, std::vector<int>& optionaction,bool destroy2) {int i;
+bool DBTestBase::CycleThroughOptions(std::vector<int>& optioncycle, std::vector<int>& optionaction,bool destroy2) {size_t i;
   for(i=0; i<optioncycle.size()-1;++i)if(option_config_==optioncycle[i])break;  // find current match
   ++i;  // advance to next option
   if(i==optioncycle.size())return false;  // if we run off the end, indicate no more options
